Reduce create_cnn code duplication
