+++ src/models/CNNDepthMap/CNNDepthMap-height/2021q1-depthmap-plaincnn-height-ensemble/src/train.py
@@ -28,9 +28,10 @@ if run.id.startswith("OfflineRun"):
     for p in utils_paths:
         shutil.copy(p, temp_model_util_dir)
 
-from model import create_cnn  # noqa: E402
 from tmp_model_util.preprocessing import preprocess_depthmap, preprocess_targets  # noqa: E402
-from tmp_model_util.utils import download_dataset, get_dataset_path, AzureLogCallback, create_tensorboard_callback, get_optimizer  # noqa: E402
+from tmp_model_util.utils import (  # noqa: E402
+    create_cnn, download_dataset, get_dataset_path, AzureLogCallback,
+    create_tensorboard_callback, get_optimizer)
 
 # Make experiment reproducible. Set random seeds for splitting.
 tf.random.set_seed(CONFIG.SPLIT_SEED)
