--- src/models/CNNDepthMap/CNNDepthMap-height/q3-depthmap-plaincnn-height/src/model.py
@@ -1,58 +1,0 @@
-from typing import Tuple
-
-import tensorflow as tf
-from tensorflow.keras import models, layers
-
-
-def create_cnn(input_shape: Tuple[int], dropout: bool) -> tf.Module:
-    model = models.Sequential()
-
-    model.add(layers.Conv2D(filters=32, kernel_size=(3, 3), padding="same", activation="relu", input_shape=input_shape))
-    model.add(layers.Conv2D(filters=32, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.05))
-
-    model.add(layers.Conv2D(filters=64, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.Conv2D(filters=64, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.075))
-
-    model.add(layers.Conv2D(filters=128, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.Conv2D(filters=128, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.1))
-
-    model.add(layers.Conv2D(filters=128, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.Conv2D(filters=128, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.125))
-
-    model.add(layers.Conv2D(filters=256, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.Conv2D(filters=256, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.15))
-
-    model.add(layers.Conv2D(filters=256, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.Conv2D(filters=256, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.175))
-
-    model.add(layers.Flatten())
-
-    model.add(layers.Dense(1024, activation="relu"))
-    if dropout:
-        model.add(layers.Dropout(0.2))
-
-    model.add(layers.Dense(128, activation="relu"))
-    if dropout:
-        model.add(layers.Dropout(0.25))
-
-    model.add(layers.Dense(1, activation="linear"))
-
-    return model
